requirements.txt
Flask
Flask-SocketIO
eventlet
python-socketio-client